+++ Assets/Scripts/UI/UpgradeButtonUI.cs
@@ -8,7 +8,7 @@ public class UpgradeButtonUI : MonoBehaviour
 {
     protected virtual int Progress { get; }
     protected virtual int MaxProgress { get; }
-    protected virtual int CostOfProgress { get; }
+    protected virtual float CostOfProgress { get; }
 
     [Space]
     [SerializeField] private Button button;
@@ -51,6 +51,7 @@ public class UpgradeButtonUI : MonoBehaviour
 
     public void UpdateText()
     {
+        /* Debug.Log(gameObject.name + ": " + CostOfProgress); */
         string txt = MoneyAmountConvertator.IntoText(CostOfProgress);
         buttonText.text = txt + "$";
